refactor(iso): move ViewRenderer into internal/view-renderer (#29)
